+++ src/lotto/validate/LottoRuleValidator.js
@@ -2,7 +2,7 @@ import {
   isLessThan,
   isGreaterThanEqualTo,
   isEqualTo
-} from '../../util/Validator.js';
+} from '../../util/ValidateUtil.js';
 import CommonValidator from './CommonValidator.js';
 import { ERROR_MESSAGE } from '../constants/Message.js';
